cranelift: Separate function generator in fuzzer
